apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "campusstudy-api"
version = "0.1.0"
description = "CampusStudy AI backend API"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "bcrypt>=4.1.3",
  "boto3>=1.35.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "pgvector>=0.3.4",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.5.2",
  "pypdf>=5.0.1",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.12",
  "python-pptx>=1.0.2",
  "PyJWT>=2.9.0",
  "redis>=5.1.1",
  "sqlalchemy>=2.0.36",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "faker>=30.3.0",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.6.9"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
